+++ app/src/main/java/com/anvob/canvasdraw/filters/action/PullInFilter.java
@@ -1,8 +1,6 @@
 package com.anvob.canvasdraw.filters.action;
 
 import android.graphics.Canvas;
-import android.graphics.Matrix;
-import android.graphics.Paint;
 import android.graphics.PorterDuff;
 import android.graphics.PorterDuffXfermode;
 
@@ -33,14 +31,14 @@ public class PullInFilter extends ActionFilter {
             int back_layer = canvas.saveLayer(0, 0, canvas.getWidth(), canvas.getHeight(), paint, Canvas.ALL_SAVE_FLAG);
             //draw transition by variants
             if (mVariant == LEFT_AND_RIGHT) {
-                int stepWidth = bitmap.getWidth() / framesCount / 2 * curFrame;
+                float stepWidth = (float) bitmap.getWidth() / framesCount / 2 * curFrame;
                 canvas.drawRect(stepWidth, 0, bitmap.getWidth() - stepWidth, bitmap.getHeight(), paint);
             } else if (mVariant == TOP_AND_DOWN) {
-                int stepHeight = bitmap.getHeight() / framesCount / 2 * curFrame;
+                float stepHeight = (float) bitmap.getHeight() / framesCount / 2 * curFrame;
                 canvas.drawRect(0, stepHeight, bitmap.getWidth(), bitmap.getHeight() - stepHeight, paint);
             } else if (mVariant == TOP_RIGHT_AND_BOTTOM_LEFT) {
-                int diag = (int) Math.sqrt(Math.pow(bitmap.getWidth(), 2) + Math.pow(bitmap.getHeight(), 2));
-                int stepDiag = diag / framesCount / 2 * curFrame;
+                float diag = (float) Math.sqrt(Math.pow(bitmap.getWidth(), 2) + Math.pow(bitmap.getHeight(), 2));
+                float stepDiag = diag / framesCount / 2 * curFrame;
                 int start = canvas.save();
                 canvas.rotate(-45, canvas.getWidth() / 2, canvas.getHeight() / 2);
                 canvas.drawRect(
@@ -50,8 +48,8 @@ public class PullInFilter extends ActionFilter {
                         bitmap.getHeight() + (diag - bitmap.getHeight()) / 2, paint);
                 canvas.restoreToCount(start);
             } else if (mVariant == TOP_LEFT_AND_BOTTOM_RIGHT) {
-                int diag = (int) Math.sqrt(Math.pow(bitmap.getWidth(), 2) + Math.pow(bitmap.getHeight(), 2));
-                int stepDiag = diag / framesCount / 2 * curFrame;
+                float diag = (float) Math.sqrt(Math.pow(bitmap.getWidth(), 2) + Math.pow(bitmap.getHeight(), 2));
+                float stepDiag = diag / framesCount / 2 * curFrame;
                 int start = canvas.save();
                 canvas.rotate(45, canvas.getWidth() / 2, canvas.getHeight() / 2);
                 canvas.drawRect(
