+++ app/src/main/java/com/anvob/canvasdraw/filters/action/PullOutFilter.java
@@ -30,14 +30,14 @@ public class PullOutFilter extends ActionFilter {
         if (curFrame < framesCount) {
             int back_layer = canvas.saveLayer(0, 0, canvas.getWidth(), canvas.getHeight(), paint, Canvas.ALL_SAVE_FLAG);
             if (getVariant() == LEFT_AND_RIGHT) {
-                int stepWidth = bitmap.getWidth() / framesCount / 2 * curFrame;
+                float stepWidth = (float) bitmap.getWidth() / framesCount / 2 * curFrame;
                 canvas.drawRect(bitmap.getWidth() / 2 - stepWidth, 0, bitmap.getWidth() / 2 + stepWidth, bitmap.getHeight(), paint);
             } else if (getVariant() == TOP_AND_DOWN) {
-                int stepHeight = bitmap.getHeight() / framesCount / 2 * curFrame;
+                float stepHeight = (float) bitmap.getHeight() / framesCount / 2 * curFrame;
                 canvas.drawRect(0, bitmap.getHeight() / 2 - stepHeight, bitmap.getWidth(), bitmap.getHeight() / 2 + stepHeight, paint);
             } else if (getVariant() == TOP_RIGHT_AND_BOTTOM_LEFT) {
-                int diag = (int) Math.sqrt(Math.pow(bitmap.getWidth(), 2) + Math.pow(bitmap.getHeight(), 2));
-                int stepDiag = diag / framesCount / 2 * curFrame;
+                float diag = (float) Math.sqrt(Math.pow(bitmap.getWidth(), 2) + Math.pow(bitmap.getHeight(), 2));
+                float stepDiag = diag / framesCount / 2 * curFrame;
                 int start = canvas.save();
                 canvas.rotate(-45, canvas.getWidth() / 2, canvas.getHeight() / 2);
                 canvas.drawRect(
@@ -47,8 +47,8 @@ public class PullOutFilter extends ActionFilter {
                         bitmap.getHeight() + (diag - bitmap.getHeight()) / 2, paint);
                 canvas.restoreToCount(start);
             } else if (getVariant() == TOP_LEFT_AND_BOTTOM_RIGHT) {
-                int diag = (int) Math.sqrt(Math.pow(bitmap.getWidth(), 2) + Math.pow(bitmap.getHeight(), 2));
-                int stepDiag = diag / framesCount / 2 * curFrame;
+                float diag = (float) Math.sqrt(Math.pow(bitmap.getWidth(), 2) + Math.pow(bitmap.getHeight(), 2));
+                float stepDiag = diag / framesCount / 2 * curFrame;
                 int start = canvas.save();
                 canvas.rotate(45, canvas.getWidth() / 2, canvas.getHeight() / 2);
                 canvas.drawRect(
