+++ app/src/main/java/com/anvob/canvasdraw/filters/action/RoundFilter.java
@@ -25,8 +25,8 @@ public class RoundFilter extends ActionFilter {
             //save current layer
             canvas.saveLayer(0, 0, canvas.getWidth(), canvas.getHeight(), paint, Canvas.ALL_SAVE_FLAG);
             //draw transition
-            int diag = (int) Math.sqrt(Math.pow(bitmap.getWidth(), 2) + Math.pow(bitmap.getHeight(), 2)) / 2;
-            int stepDiag = diag / framesCount * curFrame;
+            float diag = (float) Math.sqrt(Math.pow(bitmap.getWidth(), 2) + Math.pow(bitmap.getHeight(), 2)) / 2;
+            float stepDiag = diag / framesCount * curFrame;
 
             canvas.drawCircle(bitmap.getWidth() / 2, bitmap.getHeight() / 2, diag - stepDiag, paint);
